extract DoctrineAnnotationMatcher
